requirements.txt
asgiref==3.5.0
Django==4.0.3
django-crispy-forms==1.14.0
Pillow==9.0.1
psycopg2==2.9.3
python-decouple==3.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
validate-email==1.3
